+++ faunadb/query.py
@@ -382,9 +382,15 @@ def date(string):
 
 #region Miscellaneous functions
 
+@deprecated("use new_id instead")
 def next_id():
   """See the `docs <https://fauna.com/documentation/queries#misc_functions>`__."""
   return _fn({"next_id": None})
+
+
+def new_id():
+  """See the `docs <https://fauna.com/documentation/queries#misc_functions>`__."""
+  return _fn({"new_id": None})
 
 
 def database(db_name, scope=None):